move src to sub dir
