+++ frontend/images/browserconfig.xml
