+++ frontend/package.json
@@ -2,7 +2,7 @@
   "name": "wordbrew",
   "version": "1.0.0",
   "description": "wordbrew",
-  "main": "index.js",
+  "main": "./src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack serve --host 0.0.0.0 --mode development --stats=minimal",
